loader: move ENI rules for asymmetric routing to dedicated function
[ upstream commit 01f8dcc51c84e1cab269f84e782e09d8261ac495 ]

EKS needs some specific rules for NodePort traffic (see PR #12770, or
comments in the code, for details). The addition of part of these rules
were added to the body of the Reinitialize() function in the loader. To
make them easier to maintain or extend, let's move them to a dedicated
function called by Reinitialize(). No functional change.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>